+++ src/main.rs
@@ -1,21 +1,13 @@
-mod channel_buffers;
+pub(crate) mod channel_buffers;
+mod packet_processor;
 
-use channel_buffers::{Channel, ChannelBuffers, DeserializationError};
 use chrono::{DateTime, TimeZone, Utc};
+use packet_processor::{PacketProcessor, PacketStatistics};
 use pcap::Capture;
-use pnet::packet::{
-    ethernet::{EtherTypes, EthernetPacket},
-    ipv4::Ipv4Packet,
-    tcp::TcpPacket,
-    Packet,
-};
-use rsnano_messages::Message;
-use serde_derive::Serialize;
 use std::{
     env,
     fs::File,
     io::{BufWriter, Write},
-    net::{SocketAddr, SocketAddrV4},
 };
 
 fn main() {
@@ -26,36 +18,21 @@ fn main() {
     let mut packet_processor = PacketProcessor::new();
 
     while let Ok(packet) = capture.next() {
-        if packet_processor.statistics.packet_count % 10000 == 0 {
-            print!(".");
-            let _ = std::io::stdout().flush();
+        if packet_processor.should_indicate_progress() {
+            indicate_progress()
         }
-        packet_processor.statistics.packet_count += 1;
-        if let Some(ethernet) = EthernetPacket::new(packet.data) {
-            match ethernet.get_ethertype() {
-                EtherTypes::Ipv4 => {
-                    if let Some(ipv4) = Ipv4Packet::new(ethernet.payload()) {
-                        if let Some(tcp) = TcpPacket::new(ipv4.payload()) {
-                            let channel = into_channel(&ipv4, &tcp);
 
-                            packet_processor.add_tcp_packet(&channel, tcp.payload());
+        if let Some(channel) = packet_processor.add_packet(packet.data) {
+            let timestamp = get_timestamp(&packet);
 
-                            while let Some(entry) =
-                                packet_processor.process_packet(&channel, &packet)
-                            {
-                                serde_json::to_writer(&mut writer, &entry)
-                                    .expect("could not serialize");
-                                write!(writer, "\n").unwrap();
-                            }
-                        }
-                    }
-                }
-                _ => {}
+            while let Some(entry) = packet_processor.parse_message(&channel, timestamp) {
+                serde_json::to_writer(&mut writer, &entry).expect("could not serialize");
+                write!(writer, "\n").unwrap();
             }
         }
     }
 
-    packet_processor.statistics.print_summary();
+    print_summary(&packet_processor.statistics);
 }
 
 fn get_pcap_file_from_args() -> String {
@@ -67,105 +44,31 @@ fn get_pcap_file_from_args() -> String {
     args[1].clone()
 }
 
-pub(crate) struct PacketProcessor {
-    statistics: PacketStatistics,
-    channels: ChannelBuffers,
+fn get_timestamp(packet: &pcap::Packet) -> DateTime<Utc> {
+    Utc.timestamp_opt(
+        packet.header.ts.tv_sec,
+        packet.header.ts.tv_usec as u32 * 1000,
+    )
+    .unwrap()
 }
 
-impl PacketProcessor {
-    pub(crate) fn new() -> Self {
-        Self {
-            statistics: PacketStatistics::new(),
-            channels: ChannelBuffers::new(),
-        }
-    }
-
-    pub(crate) fn add_tcp_packet(&mut self, channel: &Channel, payload: &[u8]) {
-        self.statistics.tcp_count += 1;
-        self.channels.add_bytes(&channel, payload);
-    }
-
-    fn process_packet(&mut self, channel: &Channel, packet: &pcap::Packet) -> Option<Entry> {
-        match self.channels.pop_message(channel) {
-            Ok(None) => None,
-            Ok(Some(message)) => {
-                self.statistics.messages_parsed_count += 1;
-                Some(Entry {
-                    packet: self.statistics.packet_count,
-                    date: Utc
-                        .timestamp_opt(
-                            packet.header.ts.tv_sec,
-                            packet.header.ts.tv_usec as u32 * 1000,
-                        )
-                        .unwrap(),
-                    source: channel.source,
-                    destination: channel.destination,
-                    message,
-                })
-            }
-            Err(DeserializationError::InvalidHeader) => {
-                self.statistics.header_deserialization_failed_count += 1;
-                None
-            }
-            Err(DeserializationError::InvalidMessage) => {
-                self.statistics.message_deserialization_failed_count += 1;
-                None
-            }
-        }
-    }
+fn indicate_progress() {
+    print!(".");
+    let _ = std::io::stdout().flush();
 }
 
-#[derive(Serialize)]
-pub struct Entry {
-    packet: usize,
-    date: DateTime<Utc>,
-    source: SocketAddr,
-    destination: SocketAddr,
-    message: Message,
-}
-
-fn into_channel(ipv4: &Ipv4Packet, tcp: &TcpPacket) -> Channel {
-    Channel {
-        source: SocketAddr::V4(SocketAddrV4::new(ipv4.get_source(), tcp.get_source())),
-        destination: SocketAddr::V4(SocketAddrV4::new(
-            ipv4.get_destination(),
-            tcp.get_destination(),
-        )),
-    }
-}
-
-struct PacketStatistics {
-    packet_count: usize,
-    messages_parsed_count: usize,
-    header_deserialization_failed_count: usize,
-    message_deserialization_failed_count: usize,
-    tcp_count: usize,
-}
-
-impl PacketStatistics {
-    fn new() -> PacketStatistics {
-        PacketStatistics {
-            packet_count: 0,
-            messages_parsed_count: 0,
-            header_deserialization_failed_count: 0,
-            message_deserialization_failed_count: 0,
-            tcp_count: 0,
-        }
-    }
-
-    fn print_summary(&self) {
-        println!();
-        println!("packet_count {}", self.packet_count);
-        println!("tcp_count {}", self.tcp_count);
-        println!("messages_parsed_count {}", self.messages_parsed_count);
-        println!(
-            "header_deserialization_failed_count {}",
-            self.header_deserialization_failed_count
-        );
-        println!(
-            "message_deserialization_failed_count {}",
-            self.message_deserialization_failed_count
-        );
-        println!("Processing complete. Data written to output.csv");
-    }
+fn print_summary(stats: &PacketStatistics) {
+    println!();
+    println!("packet_count {}", stats.packet_count);
+    println!("tcp_count {}", stats.tcp_count);
+    println!("messages_parsed_count {}", stats.messages_parsed_count);
+    println!(
+        "header_deserialization_failed_count {}",
+        stats.header_deserialization_failed_count
+    );
+    println!(
+        "message_deserialization_failed_count {}",
+        stats.message_deserialization_failed_count
+    );
+    println!("Processing complete. Data written to output.csv");
 }
